metadata.ts: move in isValidURL, add processBlockContentForURLs, factored out from index.ts
